Lazily look up column names from schema during validation
Moves column name lookup from the main vlfd configuration method to a
validator, allowing the class to be loaded with validators even when the
table doesn't yet exist.  Rather than setting up Rails length validators
on each field, it registers a single validator and iterates over
columns.

This is beneficial when a model is referenced from a Rails initialiser
and loads very early, before the database is configured.  This happens
when running `rake test` as Rails generates the test database from
db/schema.rb *after* the Rails environment is initialised.

This drops support for Rails 2.x as the length validator isn't reusable.