Don't validate length on numbers
The limit value on column_schema does not specify the length in strings
but on amount of bytes it can take. It also turns out the limit numbers
can be different across different providers (they are set on mysql but
not on postgres or sqlite). Also, the length validators don't make sense
for numbers. Numbers above 2**31 (2147483648, which should correspond to
the 4 bytes integer), would pass for length validator of length 10
(such as 3000000000), even when the value would not be valid really.

Since there are no public methods in ActiveRecord to get the minimal and
maximal value, I'm rather removing the validators rather than risking
being incompatible.